requirements.txt
motor
PyJWT
bcrypt
uvicorn[standard]
pymongo
python-decouple
fastapi
pydantic[email]
passlib
pytest
jinja2
aiofiles
requests
pymongo-inmemory
httpx
